services/csm.cloud.featuretoggle.core/src/main/kotlin/com/bosch/pt/csm/cloud/featuretoggle/whitelist/command/mapper/SubjectAddedToWhitelistAvroEventMapper.kt
/*
 * ************************************************************************
 *
 *  Copyright:       Robert Bosch Power Tools GmbH, 2018 - 2022
 *
 * ************************************************************************
 */

package com.bosch.pt.csm.cloud.featuretoggle.whitelist.command.mapper

import com.bosch.pt.csm.cloud.common.model.AggregateIdentifier
import com.bosch.pt.csm.cloud.common.model.key.AggregateEventMessageKey
import com.bosch.pt.csm.cloud.featuretoggle.eventstore.FeaturetoggleAvroEventMapper
import com.bosch.pt.csm.cloud.featuretoggle.feature.messages.SubjectAddedToWhitelistEventAvro
import com.bosch.pt.csm.cloud.featuretoggle.whitelist.command.api.SubjectAddedToWhitelistEvent
import com.bosch.pt.csm.cloud.featuretogglemanagement.common.FeaturetogglemanagementAggregateTypeEnum.FEATURE_TOGGLE
import org.apache.avro.specific.SpecificRecordBase
import org.springframework.stereotype.Component

@Component
class SubjectAddedToWhitelistAvroEventMapper : FeaturetoggleAvroEventMapper {
  override fun canMap(event: Any): Boolean = event is SubjectAddedToWhitelistEvent

  override fun mapToKey(event: Any, version: Long): AggregateEventMessageKey {
    require(event is SubjectAddedToWhitelistEvent)

    return AggregateEventMessageKey(
        AggregateIdentifier(FEATURE_TOGGLE.name, event.featureId.identifier, version),
        event.featureId.identifier)
  }

  override fun mapToValue(event: Any, version: Long): SpecificRecordBase {
    require(event is SubjectAddedToWhitelistEvent)

    return SubjectAddedToWhitelistEventAvro.newBuilder()
        .apply {
          featureName = event.feature
          subjectRef = event.subjectRef.toString()
          type = event.type.name
          aggregateIdentifierBuilder = addAggregateIdentifier(event, version)
          auditingInformationBuilder = addAuditingInformation(event)
        }
        .build()
  }
}
